+++ public/index.html
@@ -13,7 +13,7 @@
   </head>
 
   <body>
-    <div style="display: none">
+    <div style="display: none;">
       <svg>
         <symbol
           id="wave"
@@ -46,6 +46,6 @@
 
       <!-- Контент -->
     </main>
-    <script type="text/javascript" src="../src/main.js"></script>
+    <script type="text/javascript" src="./bundle.js"></script>
   </body>
 </html>
