+++ src/components/load-button.js
@@ -1,0 +1,9 @@
+export const getLoadButton = () => {
+  return `
+    <button 
+      class="load-more" 
+      type="button">
+      load more
+    </button>
+  `;
+};
